externalize click off logic
